requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.2
transformers>=4.35.2
torch>=2.2.0
sentence-transformers>=2.2.2
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.1
lxml>=4.9.3
numpy>=1.24.3
scikit-learn>=1.3.2
chromadb>=0.4.15
python-dotenv>=1.0.0
